+++ src/pulse3D/transforms.py
@@ -16,21 +16,13 @@ from .constants import BESSEL_LOWPASS_10_UUID
 from .constants import BESSEL_LOWPASS_30_UUID
 from .constants import BUTTERWORTH_LOWPASS_30_UUID
 from .constants import CARDIAC_STIFFNESS_FACTOR
-from .constants import MAX_CARDIAC_EXPERIMENT_ID
-from .constants import MAX_EXPERIMENT_ID
-from .constants import MAX_SKM_EXPERIMENT_ID
-from .constants import MAX_VARIABLE_EXPERIMENT_ID
 from .constants import MICRO_TO_BASE_CONVERSION
 from .constants import MILLI_TO_BASE_CONVERSION
 from .constants import MILLIMETERS_PER_MILLITESLA
 from .constants import MILLIVOLTS_PER_MILLITESLA
-from .constants import MIN_EXPERIMENT_ID
 from .constants import NEWTONS_PER_MILLIMETER
 from .constants import RAW_TO_SIGNED_CONVERSION_VALUE
 from .constants import REFERENCE_VOLTAGE
-from .constants import ROW_LABEL_TO_VARIABLE_STIFFNESS_FACTOR
-from .constants import SKM_STIFFNESS_FACTOR
-from .constants import TWENTY_FOUR_WELL_PLATE
 from .exceptions import FilterCreationNotImplementedError
 from .exceptions import UnrecognizedFilterUuidError
 
@@ -257,18 +249,3 @@ def calculate_force_from_displacement(
     sample_in_newtons = displacement * unit_conversion * NEWTONS_PER_MILLIMETER * stiffness_factor
 
     return np.vstack((time, sample_in_newtons)).astype(np.float64)
-
-
-def get_stiffness_factor(barcode_experiment_id: int, well_idx: int) -> int:
-    if not (MIN_EXPERIMENT_ID <= barcode_experiment_id <= MAX_EXPERIMENT_ID):
-        raise ValueError(f"Experiment ID must be in the range 000-999, not {barcode_experiment_id}")
-
-    if barcode_experiment_id <= MAX_CARDIAC_EXPERIMENT_ID:
-        return CARDIAC_STIFFNESS_FACTOR
-    if barcode_experiment_id <= MAX_SKM_EXPERIMENT_ID:
-        return SKM_STIFFNESS_FACTOR
-    if barcode_experiment_id <= MAX_VARIABLE_EXPERIMENT_ID:
-        well_row_label = TWENTY_FOUR_WELL_PLATE.get_well_name_from_well_index(well_idx)[0]
-        return ROW_LABEL_TO_VARIABLE_STIFFNESS_FACTOR[well_row_label]
-    # if experiment ID does not have a stiffness factor defined (currently 300-999) then just use the value for Cardiac
-    return CARDIAC_STIFFNESS_FACTOR